Workstation should manually cache and invalidate audio in memory as needed by craft/dub ahead https://www.pivotaltracker.com/story/show/186556102

+++ nexus/src/main/java/io/xj/nexus/work/WorkConfiguration.java
@@ -17,7 +17,7 @@ public class WorkConfiguration {
   private int outputSeconds;
   private long craftAheadMicros;
   private long dubAheadMicros;
-  private double outputFrameRate;
+  private int outputFrameRate;
   private int outputChannels;
   private String contentStoragePathPrefix;
   private int mixerLengthSeconds = 2;
@@ -111,11 +111,11 @@ public class WorkConfiguration {
     return this;
   }
 
-  public double getOutputFrameRate() {
+  public int getOutputFrameRate() {
     return outputFrameRate;
   }
 
-  public WorkConfiguration setOutputFrameRate(double outputFrameRate) {
+  public WorkConfiguration setOutputFrameRate(int outputFrameRate) {
     this.outputFrameRate = outputFrameRate;
     return this;
   }
